+++ fugassem/workflows/fugassem_main.py
@@ -109,6 +109,9 @@ def parse_cli_arguments ():
 	workflow.add_argument("matrix-list",
 	                      desc = "[string] a comma separated list of matrix-based evidence files (e.g. gene-by-gene network type of file) [ Default: None ]",
 	                      default = None)
+	workflow.add_argument("matrix-pair",
+	                      desc = "treat matrix-based evidence as pair-based network", 
+	                      action = "store_true")
 	workflow.add_argument("basename",
 	                      desc="specify the basename for output files [ Default: fugassem ]",
 	                      default = "fugassem")
@@ -271,6 +274,11 @@ def fugassem_main (workflow):
 			args.minimum_number = 10
 	else:
 		args.minimum_number = 10
+	
+	if args.matrix_pair:
+		pair_flag = "yes"
+	else:
+		pair_flag = "no"
 
 	## get all input files
 	go_obo = config.go_obo
@@ -351,14 +359,15 @@ def fugassem_main (workflow):
 		               args.memory,
 		               args.time,
 		               myoutput_dir,
-		               mylog]
+		               mylog,
+					   pair_flag]
 			target_list = [final_func_file, final_func_smp_file, final_funclist_file, feature_list_file, final_pred_file]
 			workflow.add_task_gridable(
 				"fugassem_process --input [depends[0]] --gene [depends[1]] --function [depends[2]] "
 				"--taxon [args[0]] --basename [args[1]] "
 				"--minimum-prevalence [args[2]] --minimum-abundance [args[3]] --minimum-detected [args[4]] --filtering-zero [args[5]] --covariate-taxon [args[6]] "
 				"--correlation-method [args[7]] --go-level [args[8]] --func-type [args[9]] --ml-type [args[10]] "
-				"--vector-list [args[11]] --matrix-list [args[12]] "
+				"--vector-list [args[11]] --matrix-list [args[12]] --pair-flag [args[21]] "
 				"--bypass-preprocessing [args[13]] --bypass-prediction [args[14]] --bypass-mtx [args[15]] "
 				"--threads [args[16]] --memory [args[17]] --time [args[18]] "
 				"--output [args[19]] > [args[20]] 2>&1",