+++ history.md
@@ -3,6 +3,7 @@
 
 ## v0.3.0 2022-02-05 ##
 * Accepted input matrix evidence with connection weights
+* Added options to accept either pair- or unpair-based matrix evidence
 
 ## v0.2.1 2022-01-08 ##
 * Fixed issues of utilities for preparing input files